Re-use DbHelper between the two MySql implementations, like the statement generator
